Fix: Modal background issue 해결
