+++ src/components/Post/PostList/PostList.jsx
@@ -6,7 +6,8 @@ import IconAlbumOn from "../../../assets/images/icon-post-album-on.svg";
 import IconListOff from "../../../assets/images/icon-post-list-off.svg";
 import IconListOn from "../../../assets/images/icon-post-list-on.svg";
 import axios from "axios";
-import { useLocation, Link, useNavigate } from "react-router-dom";
+import { useLocation, useNavigate } from "react-router-dom";
+import Modal from "../../Modal/Modal";
 
 const PostListDiv = styled.div`
   display: flex;
@@ -143,6 +144,20 @@ export default function PostList({ post, modalOpen }) {
     });
   }
 
+  const [modalShow, setModalShow] = useState(false);
+  const [selectedId, setSelectedId] = useState(null);
+
+  function modalClose(e) {
+    if (e.target === e.currentTarget) {
+      setModalShow(false);
+    }
+  }
+
+  function modalOpen(id) {
+    setSelectedId(id);
+    setModalShow(true);
+  }
+
   return (
     <>
       {hasPosts && (
@@ -193,6 +208,13 @@ export default function PostList({ post, modalOpen }) {
             </GridItemList>
           )}
         </>
+      )}
+      {modalShow && (
+        <Modal
+          type="modification"
+          modalClose={modalClose}
+          postId={selectedId}
+        />
       )}
     </>
   );
